controller admin

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -19,14 +19,20 @@ class CategoryController extends Controller
         if ($user->is_blocked) {
             return view('blocked.index');
         } else {
-            $data = DB::table('categories as c')
-                ->leftJoin('products as p', 'c.id', '=', 'p.category_id')
-                ->select('c.id', 'c.name', DB::raw('COUNT(p.id) as count'))
-                ->groupBy('c.id', 'c.name')
-                ->orderByDesc('id')
-                ->get();
-            return view('admin.category.categories', ['data' => $data]);
+            if ($user->is_admin) {
+                $data = DB::table('categories as c')
+                    ->leftJoin('products as p', 'c.id', '=', 'p.category_id')
+                    ->select('c.id', 'c.name', DB::raw('COUNT(p.id) as count'))
+                    ->groupBy('c.id', 'c.name')
+                    ->orderByDesc('id')
+                    ->get();
+                return view('admin.category.categories', ['data' => $data]);
+            } else {
+                return redirect()->route("client.produit.index");
+            }
         }
+
+
     }
 
     /**
@@ -46,16 +52,22 @@ class CategoryController extends Controller
         if ($user->is_blocked) {
             return view('blocked.index');
         } else {
-            $data['name'] = $request->name;
+            if ($user->is_admin) {
+                // ------------------------------
+                $data['name'] = $request->name;
 
-            $check = Category::where('name', $request->name)->first();
-            if ($check) {
-                return redirect()->route("admin.categories.index")->with('danger', '"' . $check->name . '" already exist.');
+                $check = Category::where('name', $request->name)->first();
+                if ($check) {
+                    return redirect()->route("admin.categories.index")->with('danger', '"' . $check->name . '" already exist.');
+                }
+
+                $data['created_at'] = date('Y-m-d H:i:s');
+                Category::create($data);
+                return redirect()->route("admin.categories.index")->with('success', 'category has been created successfully.');
+
+            } else {
+                return redirect()->route("client.produit.index");
             }
-
-            $data['created_at'] = date('Y-m-d H:i:s');
-            Category::create($data);
-            return redirect()->route("admin.categories.index")->with('success', 'category has been created successfully.');
         }
     }
 
@@ -76,8 +88,13 @@ class CategoryController extends Controller
         if ($user->is_blocked) {
             return view('blocked.index');
         } else {
-            $data = Category::select('*')->find($id);
-            return view("admin.category.editcategories", ['data' => $data]);
+            if ($user->is_admin) {
+                // ------------------------------
+                $data = Category::select('*')->find($id);
+                return view("admin.category.editcategories", ['data' => $data]);
+            } else {
+                return redirect()->route("client.produit.index");
+            }
         }
     }
 
@@ -90,18 +107,25 @@ class CategoryController extends Controller
         if ($user->is_blocked) {
             return view('blocked.index');
         } else {
-            $data['name'] = $request->name;
+            if ($user->is_admin) {
+                // ------------------------------
+                $data['name'] = $request->name;
 
-            $check = Category::where('name', $request->name)->whereNotIn('id', [$id])->first();
-            if ($check) {
-                return redirect()->route("admin.categories.edit", $id)->with('danger', '"' . $check->name . '" already exist.');
+                $check = Category::where('name', $request->name)->whereNotIn('id', [$id])->first();
+                if ($check) {
+                    return redirect()->route("admin.categories.edit", $id)->with('danger', '"' . $check->name . '" already exist.');
+                }
+
+                $data['name'] = $request->name;
+                $data['updated_at'] = date('Y-m-d H:i:s');
+                Category::where('id', $id)->update($data);
+                return redirect()->route("admin.categories.index")->with('success', 'category has been edited successfully.');
+
+            } else {
+                return redirect()->route("client.produit.index");
             }
-
-            $data['name'] = $request->name;
-            $data['updated_at'] = date('Y-m-d H:i:s');
-            Category::where('id', $id)->update($data);
-            return redirect()->route("admin.categories.index")->with('success', 'category has been edited successfully.');
         }
+
     }
 
     /**
@@ -113,8 +137,13 @@ class CategoryController extends Controller
         if ($user->is_blocked) {
             return view('blocked.index');
         } else {
-            Category::where('id', $id)->delete();
-            return redirect()->route('admin.categories.index')->with('success', 'category has been deleted successfully.');
+            if ($user->is_admin) {
+                // ------------------------------
+                Category::where('id', $id)->delete();
+                return redirect()->route('admin.categories.index')->with('success', 'category has been deleted successfully.');
+            } else {
+                return redirect()->route("client.produit.index");
+            }
         }
     }
 }
